fixed types of api parameters.

+++ XSLTViz/DataModel/Point.cs
@@ -11,9 +11,9 @@ namespace XSLTViz.DataModel
     public class Point
     {
         [JsonProperty(PropertyName = "x")]
-        public int? X { get; set; }
+        public double? X { get; set; }
 
         [JsonProperty(PropertyName = "y")]
-        public int? Y { get; set; }
+        public double? Y { get; set; }
     }
 }
